Added functionality: Highlight text from article and create new person from highlighted text.
updatePersonNameFromHighlight() reads the highlighted text from the article, splits it, and sets the values of the 'new-person' modal. This function calls checkPrefix() and checkSuffix() functions.

Added Functionality: Added 'addPersonFactMultiple' fact button to the top right corner of screen. This button allows the user to select multiple persons, fill in fact information, and add this fact to all of the selected persons.

'addPersonFactMultiple' gets the id of 'new-relationship-person-3' which contains the list of all persons and loops through all selected names. After the loop it updates the editor.